Fix mean / std estimation for psf
