+++ pyirf/interpolation/base_interpolators.py
@@ -3,9 +3,11 @@ from abc import ABCMeta, abstractmethod
 import enum
 
 import numpy as np
-from pyirf.binning import bin_center
 
-__all__ = ["BaseInterpolator", "ParametrizedInterpolator", "DiscretePDFInterpolator"]
+from ..binning import bin_center
+from ..utils import cone_solid_angle
+
+__all__ = ["BaseInterpolator", "ParametrizedInterpolator", "DiscretePDFInterpolator", "PDFNormalization", "get_bin_width"]
 
 
 class PDFNormalization(enum.Enum):
@@ -15,6 +17,16 @@ class PDFNormalization(enum.Enum):
     #: PDF is normalized to 1 over the solid angle integral where the bin
     #: edges represent the opening angles of cones in radean.
     CONE_SOLID_ANGLE = enum.auto()
+
+
+def get_bin_width(bin_edges, normalization):
+    if normalization is PDFNormalization.AREA:
+        return np.diff(bin_edges)
+
+    if normalization is PDFNormalization.CONE_SOLID_ANGLE:
+        return np.diff(cone_solid_angle(bin_edges))
+    
+    raise ValueError(f"Invalid PDF normalization: {normalization}")
 
 
 class BaseInterpolator(metaclass=ABCMeta):
